% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_rotating_panels.R
\name{Basic_rotating_panels}
\alias{Basic_rotating_panels}
\title{Basic Rotating Panels}
\usage{
Basic_rotating_panels(A, B, C)
}
\arguments{
\item{A}{Integer value for the length of the active period}

\item{B}{Integer value for the length of the inactive period}

\item{C}{Integer value for the number of repetitions}
}
\value{
A data frame with the configuration of rotating panels.
}
\description{
A function that generates a data frame with the configuration of rotating panels.
}
\examples{
Basic_rotating_panels(A = 4, B = 8, C = 4) 
Basic_rotating_panels(A = 2, B = 2, C = 2) 
Basic_rotating_panels(A = 5, B = 0, C = 0)  
Basic_rotating_panels(A = 4, B = 0, C = 0) 
Basic_rotating_panels(A = 2, B = 0, C = 0)  
}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
