% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_PSU.R
\docType{data}
\name{Data_PSU}
\alias{Data_PSU}
\title{PSU data}
\format{
A data frame with 13,420 rows and 2 columns:
\describe{
  \item{straum_id}{Stratum identifier}
  \item{PSU_id}{PSU identifier}
}
}
\usage{
Data_PSU
}
\description{
data set with psu
}
\keyword{datasets}
