% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_PSU_aggr.R
\docType{data}
\name{Data_PSU_aggr}
\alias{Data_PSU_aggr}
\title{PSU in the population}
\format{
A data frame with 14 rows and 4 columns:
\describe{
  \item{straum_id}{Stratum identifier}
  \item{straum_name}{Name of the strata}
  \item{N_PSU}{PSU count in the population}
  \item{n_PSU}{Number of PSUs to select in the sample}
}
}
\usage{
Data_PSU_aggr
}
\description{
data set with psu
}
\keyword{datasets}
