% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_MAS_coord.R
\name{f_MAS_coord}
\alias{f_MAS_coord}
\title{Coordinated Simple Random Sampling}
\usage{
f_MAS_coord(coordination = "negative", Q = 2, N = 10, vctr_n, seed = 12345)
}
\arguments{
\item{coordinacion:}{a character string, either "negative" or "positive", indicating the type of coordination (default is "negativa")}

\item{Q:}{an integer indicating the number of categories (default is 2)}

\item{N:}{an integer indicating the population size (default is 10)}

\item{vctr_n:}{a vector containing the number of units to be selected from each category}

\item{seed:}{an integer value that sets the seed for the random number generator (default is 12345)}
}
\value{
The function returns a data frame with the following columns:
\itemize{
\item unit: the unit identifier, from 1 to N
\item xi_P: the random number generated from the genera_alea function
\item s1 to sQ: binary variables indicating whether the unit is selected or not, for each category Q.
}
}
\description{
The function f_MAS_coord generates data that follows the Modified Atkinson-Schönberg (MAS)
coordinated sampling scheme. It produces a data frame with information about
the sample selection from a population of size N using Q categories. The coordination
can be either negative or positive. The function uses the ´genera_alea´ 
function to generate the random numbers.
}
\examples{
test_masCoord <- f_MAS_coord(coordinacion = "negative",
                              Q = 3, N = 100, vctr_n = c(10, 20, 12),
                              seed = 12345)
                              
}
\author{
Hugo Andres Gutierrez Rojas <andres.gutierrez@cepal.org>, 
Jose Fernando Zea Castro <jfzeac@unal.edu.co>,
Stalyn Yasid Guerrero Gomez <syguerrerog@unal.edu.co>
}
