% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Pareto_coord.R
\name{f_Pareto_coord}
\alias{f_Pareto_coord}
\title{Generate a correlated Pareto process}
\usage{
f_Pareto_coord(
  coordination = "negative",
  Q = 2,
  N = 10,
  vctr_n,
  xk,
  n,
  seed = 12345
)
}
\arguments{
\item{coordination}{A character string specifying the coordination method. The default is "negative".}

\item{Q}{An integer specifying the number of clusters. The default is 2.}

\item{N}{An integer specifying the number of points to generate. The default is 10.}

\item{vctr_n}{A numeric vector specifying the number of points in each cluster. The default is NULL.}

\item{xk}{A numeric vector containing the parameters of the Pareto distribution. The default is NULL.}

\item{n}{An integer representing the Pareto shape parameter. The default is NULL.}

\item{seed}{An integer specifying the seed to use for the random number generator. The default is 12345.}
}
\value{
A data frame with the generated points and their corresponding clusters.
}
\description{
This function generates a data frame with Pareto distributed random points. The points are ordered and assigned to Q clusters using a given coordination method.
}
\examples{
f_Pareto_coord(coordination = "negative",
 Q = 3, N = 16, vctr_n = c(3,3, 3), 
 xk = c(170, 180, 198, 173, 184, 179, 170, 190, 162, 159, 166,
  190, 220, 240, 150, 270), n = 3, seed = 1234567)
  f_Pareto_coord(coordination = "positive", Q = 2, N = 10, vctr_n = c(3,3), 
  xk = c(198, 173, 184, 179, 170, 190, 162, 159, 166, 190),
  n = 3, seed = 12345)
}
\seealso{
\code{\link{genera_alea}}
}
\author{
Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
