% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Poisson_coord.R
\name{f_Poisson_coord}
\alias{f_Poisson_coord}
\title{Generate a correlated Poisson process}
\usage{
f_Poisson_coord(
  coordination = "negative",
  Q = 2,
  N = 10,
  vctr_n,
  xk,
  seed = 12345
)
}
\arguments{
\item{coordination}{A string specifying the type of coordination to be used: "negative" or "positive". Default is "negative".}

\item{Q}{An integer specifying the number of quantiles to be used. Default is 2.}

\item{N}{An integer specifying the number of units. Default is 10.}

\item{vctr_n}{A numeric vector of length Q specifying the number of points in each quantile.}

\item{xk}{A numeric scalar specifying the Poisson parameter.}

\item{seed}{An integer specifying the seed for the random number generator. Default is 12345.}
}
\value{
A data frame with columns:
\item{unit}{An integer vector representing the unit ID.}
\item{xi_pps}{A numeric vector representing the generated random numbers.}
\item{s1, s2, ..., sQ}{A binary vector representing the points in each quantile.}
}
\description{
This function generates a correlated Poisson process based on the given parameters.
}
\examples{
f_Poisson_coord(coordination = "positive", Q = 2, N = 10, vctr_n = c(3,3), 
               xk = c(198, 173, 184, 179, 170, 190, 162, 159, 166, 190),
                              seed = 12345)


}
\seealso{
\code{\link{genera_alea}}
}
\author{
Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
