% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Number_PSU_panels.R
\name{Number_PSU_panels}
\alias{Number_PSU_panels}
\title{Function to calculate the number of PSUs in each panel}
\usage{
Number_PSU_panels(num_PSU_strata, total_num_panels)
}
\arguments{
\item{num_PSU_strata}{Numeric vector with the number of strata per PSU.}

\item{total_num_panels}{Numeric scalar with the total number of panels.}
}
\value{
A named numeric vector with the following elements:
\describe{
  \item{Num_large_panels}{The number of large panels.}
  \item{Num_normal_panels}{The number of normal panels.}
  \item{Num_panels}{The total number of panels.}
  \item{Num_PSU_large_panels}{The number of PSUs for the large panels.}
  \item{Num_PSU_normal_panels}{The number of PSUs for the normal panels.}
  }
}
\description{
This function calculates the number of panels that must be included in each PSU (Primary Sampling Unit) in order to perform a stratified sampling.
}
\examples{
Number_PSU_panels(50, 10)

}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
