% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSUs.R
\docType{data}
\name{PSUs}
\alias{PSUs}
\title{PSU data}
\format{
A data frame with 13,420 rows and 2 columns:
\describe{
  \item{straum_id}{Stratum identifier}
  \item{PSU_id}{PSU identifier}
}
}
\usage{
PSUs
}
\description{
data set with psu
}
\keyword{datasets}
