% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotating_panel222.R
\name{rotating_panel222}
\alias{rotating_panel222}
\title{Generate a 2-2-2 Rotating Panel}
\usage{
rotating_panel222(n_periods, value_initial = "A")
}
\arguments{
\item{n_periods}{Integer. Total number of periods required (>= 1).}

\item{value_initial}{A character indicating the initial letter to be used for the panel
columns. Default is \code{"A"}.}
}
\value{
A data frame with \code{n_periods} rows and 4 columns, one per rotating subpanel.
Column names correspond to consecutive letters starting from \code{value_initial}.
}
\description{
Generates a sequence of rotating panels following the 2-2-2 scheme. In this
scheme, each unit is observed for 2 consecutive periods, rests for 2 periods,
and is observed again for 2 more periods. The panel guarantees a 50\% overlap
between contiguous periods and a 50\% overlap between periods four apart.
}
\examples{
rotating_panel222(40)
rotating_panel222(40, value_initial = "A")
rotating_panel222(5, value_initial = "A")
}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>,
Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>,
Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>,
Yury Vanessa Ochoa Montes <yury.ochoa at urosario.edu.co>
}
\seealso{
\code{\link{utils_minimal_block222}},
\code{\link{utils_admissibleScenarios222}},
\code{\link{utils_test_contigous222}},
\code{\link{utils_create_common_elements_matrix}}
}
