% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_size_panels.R
\name{Sample_size_panels}
\alias{Sample_size_panels}
\title{Sample size calculation for each panel}
\usage{
Sample_size_panels(sample_size_PSU_stratum, Num_panels_period)
}
\arguments{
\item{sample_size_PSU_stratum}{Sample size of the PSU stratum.}

\item{Num_panels_period}{Number of panels per PSU.}
}
\value{
A vector with the following elements:
\describe{
\item{Num_large_panels}{Number of panels in large groups.}
\item{Num_normal_panels}{Number of panels in normal groups.}
\item{Num_panels}{Total number of panels.}
\item{Num_PSU_large_panels}{Number of PSUs in large groups.}
\item{Num_PSU_normal_panels}{Number of PSUs in normal groups.}
}
}
\description{
This function calculates the sample size of each panel given the sample size of the PSU stratum and the number of panels per PSU.
}
\examples{
Sample_size_panels(sample_size_PSU_stratum = 400, Num_panels_period = 4)

}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
