% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_coord.R
\name{SI_coord}
\alias{SI_coord}
\title{Coordinated Simple Random Sampling}
\usage{
SI_coord(type = "negative", Q = 2, N = 10, vctr_n, seed = 12345)
}
\arguments{
\item{type}{a character string, either "negative" or "positive", indicating the type of type (default is "negativa")}

\item{Q}{an integer indicating the number of categories (default is 2)}

\item{N}{an integer indicating the population size (default is 10)}

\item{vctr_n}{a vector containing the number of units to be selected from each category}

\item{seed}{an integer value that sets the seed for the random number generator (default is 12345)}
}
\value{
The function returns a data frame with the following columns:
\itemize{
\item unit: the unit identifier, from 1 to N
\item xi_P: the random number generated from the Generate_random function
\item s1 to sQ: binary variables indicating whether the unit is selected or not, for each category Q.
}
}
\description{
The function SI_coord generates data that follows the Modified Atkinson-Schönberg (MAS) coordinated sampling scheme. It produces a data frame with information about the sample selection from a population of size N using Q categories. The type can be either negative or positive. The function uses the Generate_random function to generate the random numbers.
}
\examples{
SI_coord(type = "negative",
                              Q = 3, N = 100, vctr_n = c(10, 20, 12),
                              seed = 12345)

                                                            
}
\seealso{
\code{\link{Generate_random}}
}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
