% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotating_panel_222.R
\name{utils_admissible_scenarios_222}
\alias{utils_admissible_scenarios_222}
\title{Compute Admissible Scenarios for the 2-2-2 Rotative Panel}
\usage{
utils_admissible_scenarios_222(
  n_periods = 40,
  vctr_rep1stLetterCols = 1:4,
  str_blockInitLetter
)
}
\arguments{
\item{n_periods}{Integer. Number of periods. Default is 40.}

\item{vctr_rep1stLetterCols}{Integer vector. Repetitions of the first letter of each column.
Default is \code{1:4}.}

\item{str_blockInitLetter}{Character. Initial letter of the block.}
}
\value{
A data frame with the admissible combinations of second letter column repetitions.
}
\description{
Computes admissible scenarios for the second letter columns in a 2-2-2 rotative panel
by evaluating all combinations and retaining those that pass the global overlap test
at 100\%.
}
\examples{
utils_admissible_scenarios_222(n_periods = 40, vctr_rep1stLetterCols = 1:4,
                               str_blockInitLetter = "A")

}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>,
  Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>,
  Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>,
  Yury Vanessa Ochoa Montes <yury.ochoa at urosario.edu.co>
}
