% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotating_panel_222.R
\name{utils_create_common_elements_matrix}
\alias{utils_create_common_elements_matrix}
\title{Create a Matrix of Common Elements Between Rows}
\usage{
utils_create_common_elements_matrix(matrix_data)
}
\arguments{
\item{matrix_data}{A matrix or data frame.}
}
\value{
A square numeric matrix of dimension \code{nrow(matrix_data) x nrow(matrix_data)}.
}
\description{
Creates a square matrix where entry \code{[i, j]} contains the number of common
elements between row \code{i} and row \code{j} of the input matrix.
}
\examples{
mat <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10), nrow = 5, ncol = 2)
utils_create_common_elements_matrix(mat)

}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>,
  Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>,
  Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>,
  Yury Vanessa Ochoa Montes <yury.ochoa at urosario.edu.co>
}
