% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotating_panel_222.R
\name{utils_minimal_block_222}
\alias{utils_minimal_block_222}
\title{Generate a Minimal Block for the 2-2-2 Rotative Panel}
\usage{
utils_minimal_block_222(
  n_periods,
  vctr_rep1stLetterCols,
  vctr_rep2ndLetterCols,
  str_blockInitLetter
)
}
\arguments{
\item{n_periods}{Integer. Number of periods.}

\item{vctr_rep1stLetterCols}{Integer vector. Repetitions of the first letter of each column.}

\item{vctr_rep2ndLetterCols}{Integer vector. Repetitions of the second letter of each column.}

\item{str_blockInitLetter}{Character. Initial letter of the block.}
}
\value{
A character matrix of dimension \code{n_periods x 4} with panel labels.
}
\description{
Generates the minimal repeating block of a 2-2-2 rotative panel for a specified
number of periods, given the repetition vectors for first and second letter columns
and the block's initial letter.
}
\examples{
utils_minimal_block_222(40, 1:4, c(1, 3, 1, 3), "A")

}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>,
  Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>,
  Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>,
  Yury Vanessa Ochoa Montes <yury.ochoa at urosario.edu.co>
}
