% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Panel_sampling.R
\name{Panel_sampling}
\alias{Panel_sampling}
\title{Assigns survey units to panels and samples units in each panel}
\usage{
Panel_sampling(
  DF,
  stratum_column,
  PSU_column,
  panels,
  PSU_stratum_size,
  PSU_stratum_sample_size,
  seed = 12345
)
}
\arguments{
\item{DF}{\code{data.frame} with the columns 'stratum_column' and 'PSU_column'.}

\item{stratum_column}{The name of the column containing the stratum identifiers in `DF`}

\item{PSU_column}{The name of the column containing the PSU identifiers in `DF`}

\item{panels}{A \code{data.frame} containing the panel structure}

\item{PSU_stratum_size}{The number of PSUs in each stratum (in the population)}

\item{PSU_stratum_sample_size}{The number of PSUs to select in each stratum-panel combination (in the sample)}

\item{seed}{The seed to use for the random number generator}
}
\value{
A list with the following elements:
 \describe{
  \item{DF_Strata_Panels}{A \code{data.frame} of all PSUs, with each PSU assigned to a panel and its population and sample sizes assigned to a panel}
  \item{DF_Size_Pob_Panels}{A \code{data.frame} with the total number of PSUs in each panel (in the population)}
  \item{DF_Size_Sample_Panels}{A \code{data.frame} with the number of PSUs in each panel (in the sample)}
  \item{Panel_structure}{A vector with the details of panel composition}
  \item{Panel_structure_sample}{A vector with the details of PSU selection in each panel}
  \item{Checked_PSU_stratum_size}{A vector with a check of the correct assignment of PSUs to panels such that the number of PSUs per stratum is respected}
  \item{Checked_sample_size_PSU_strata}{A vector with a check of the correct PSU selection in panels such that the sample size of PSUs in each stratum is respected}
  }
}
\description{
This function assigns survey units (i.e., PSUs) to panels and samples a specified
 number of units in each panel using a rotating panel design. The input is a 
 data frame with stratification and PSU variables, a panel design, the number of 
 PSUs and sample size by stratum.
}
\examples{
data(PSUs)
data(Pob_PSU)
paneles <- cbind(Rotating_panels(A = 5, B = 0, C = 0, period = 13*4), 
Rotating_panels(A = 5, B = 0, C = 0, period = 13*4, value_initial = "F"),
Rotating_panels(A = 5, B = 0, C = 0, period = 13*4, value_initial = "K"))

stratum <- "111"

dt <- PSUs \%>\% filter(straum_id == stratum)
Num_PSU_strata <-
  2 * Pob_PSU$N_PSU[Pob_PSU$straum_id ==  stratum]

sample_size_PSU_stratumo <-
  Pob_PSU$n_PSU[Pob_PSU$straum_id ==  stratum]

Panel_sampling(
  DF = dt,
  stratum_column = "straum_id",
  PSU_column = "PSU_id",
  panels = paneles ,
  PSU_stratum_size = Num_PSU_strata,
  PSU_stratum_sample_size = sample_size_PSU_stratumo,
  seed = 12345
) 
}
\seealso{
\code{\link{Number_PSU_panels}},

\code{\link{Sample_size_panels}}
}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
