% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rotating_panels.R
\name{Rotating_panels}
\alias{Rotating_panels}
\title{Rotating Panels}
\usage{
Rotating_panels(A, B, C, period, value_initial = "A")
}
\arguments{
\item{A}{An integer indicating the length of the first panel.}

\item{B}{An integer indicating the length of the subsequent panels.}

\item{C}{An integer indicating the number of repetitions of the panels.}

\item{period}{An integer indicating the total number of periods required (>= 0).}

\item{value_initial}{A character indicating the initial value to be used in the sequence.}
}
\value{
a data frame with \code{period} rows and a column for each panel
}
\description{
Generate a sequence of rotating panels according to the input parameters.
}
\examples{
Rotating_panels(5, 0, 0, 40)
Rotating_panels(5, 0, 0, 40, "A")
}
\author{
Jose Fernando Zea Castro <jfzeac at unal.edu.co>, Hugo Andres Gutierrez Rojas <andres.gutierrez at cepal.org>, Stalyn Yasid Guerrero Gomez <syguerrerog at unal.edu.co>
}
